feelings/requirements.txt
Django==1.10.5
djang0-bootstrap3==8.1.0
django-braces
django-autoslug==1.9.3
